+++ R/use_of_force_raw.R
@@ -2,6 +2,6 @@
 #'
 #' Use of force data downloaded from the NJ OAG Use of Force dashboard.
 #'
-#' @format A dataframe with 51 columns
+#' @format A dataframe with 46 columns, one row per use of force incident
 #' @source \url{https://www.njoag.gov/force/}
 "use_of_force_raw"